+++ examples/confirm.rs
@@ -1,6 +1,17 @@
 use dialoguer::Confirm;
 
 fn main() {
+    if Confirm::new()
+        .with_prompt("Do you want to continue?")
+        .interact()
+        .unwrap()
+    {
+        println!("Looks like you want to continue");
+    } else {
+        println!("nevermind then :(");
+        return;
+    }
+
     println!("disable default");
     if Confirm::new()
         .with_prompt("continue?")