+++ examples/password.rs
@@ -6,5 +6,6 @@ fn main() {
         .with_confirmation("Repeat password", "Error: the passwords don't match.")
         .interact()
         .unwrap();
+
     println!("Your password is {} characters long", password.len());
 }